requirements.txt
faster-whisper==1.0.3
torch>=2.1.0
torchaudio>=2.1.0
numpy>=1.26.0
PyYAML>=6.0
requests>=2.31.0
Pillow>=10.0.0
ffmpeg-python>=0.2.0
